requirements.txt
# 掌静脉识别系统依赖项
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0
opencv-python>=4.4.0
scipy>=1.5.0
tqdm>=4.50.0
# 如果需要使用pytorch2caffe转换工具，还需要安装以下依赖
# caffe
pytorch2caffe